+++ src/main/java/com/wynntils/modules/core/overlays/inventories/InventoryReplacer.java
@@ -7,6 +7,8 @@ package com.wynntils.modules.core.overlays.inventories;
 import com.wynntils.core.events.custom.GuiOverlapEvent;
 import com.wynntils.core.framework.FrameworkManager;
 import com.wynntils.modules.questbook.QuestBookModule;
+import com.wynntils.modules.questbook.managers.QuestBookHandler;
+import com.wynntils.modules.questbook.overlays.ui.MainPage;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.gui.inventory.GuiInventory;
 import net.minecraft.entity.player.EntityPlayer;
@@ -59,7 +61,7 @@ public class InventoryReplacer extends GuiInventory {
     @Override
     public void actionPerformed(GuiButton guiButton) throws IOException {
         if (guiButton.id == 10) {
-            QuestBookModule.gui.open();
+            QuestBookHandler.openQuestBookPage(true, MainPage.class);
             return;
         }
