+++ src/main/java/com/wynntils/modules/questbook/QuestBookModule.java
@@ -9,19 +9,13 @@ import com.wynntils.core.framework.instances.Module;
 import com.wynntils.core.framework.interfaces.annotations.ModuleInfo;
 import com.wynntils.modules.questbook.configs.QuestBookConfig;
 import com.wynntils.modules.questbook.events.ClientEvents;
-import com.wynntils.modules.questbook.managers.QuestManager;
 import com.wynntils.modules.questbook.overlays.hud.TrackedQuestOverlay;
-import com.wynntils.modules.questbook.overlays.ui.MainPage;
-import com.wynntils.modules.questbook.overlays.ui.QuestBookGUIold;
+import com.wynntils.modules.questbook.overlays.ui.*;
 import com.wynntils.modules.questbook.managers.QuestBookHandler;
-import com.wynntils.modules.questbook.overlays.ui.QuestsPage;
-import com.wynntils.modules.questbook.overlays.ui.SettingsPage;
 import org.lwjgl.input.Keyboard;
 
 @ModuleInfo(name = "quest_book", displayName = "Quest Book")
 public class QuestBookModule extends Module {
-
-    public static final QuestBookGUIold gui = new QuestBookGUIold();
 
     public void onEnable() {
         registerEvents(new ClientEvents());
@@ -32,11 +26,16 @@ public class QuestBookModule extends Module {
         QuestBookHandler.registerPage(new MainPage());
         QuestBookHandler.registerPage(new QuestsPage());
         QuestBookHandler.registerPage(new SettingsPage());
+        QuestBookHandler.registerPage(new ItemPage());
+        QuestBookHandler.registerPage(new HUDConfigPage());
+        QuestBookHandler.registerPage(new DiscoveriesPage());
 
         registerKeyBinding("Open Quest Book", Keyboard.KEY_K, "Wynntils", true, () -> {
             QuestBookHandler.openQuestBookPage(true, QuestsPage.class);
         });
-        registerKeyBinding("Open Item Guide", Keyboard.KEY_I, "Wynntils", true, gui::openAtItemGuide);
+        registerKeyBinding("Open Item Guide", Keyboard.KEY_I, "Wynntils", true, () -> {
+            QuestBookHandler.openQuestBookPage(true, MainPage.class); }
+            );
     }
 
 }
