--- src/main/java/com/wynntils/modules/questbook/enums/QuestBookPage.java
@@ -1,11 +1,0 @@
-/*
- *  * Copyright © Wynntils - 2019.
- */
-
-package com.wynntils.modules.questbook.enums;
-
-public enum QuestBookPage {
-
-    DEFAULT, QUESTS, CONFIGS, ITEM_GUIDE, DISCOVERIES
-
-}
