+++ src/main/java/com/wynntils/modules/questbook/instances/IconContainer.java
@@ -2,18 +2,21 @@ package com.wynntils.modules.questbook.instances;
 
 public enum IconContainer {
 
-    questPageIcon(0, 24, 221, 255),
-    settingsPageIcon(283, 304, 221, 263),
-    itemGuideIcon(307, 325, 221, 261),
-    hudConfigIcon(262, 283, 261, 303);
+    questPageIcon(0, 24, 221, 255, true),
+    settingsPageIcon(283, 304, 221, 263, true),
+    itemGuideIcon(307, 325, 221, 261, true),
+    hudConfigIcon(262, 283, 261, 303, true),
+    discoveriesIcon(206, 221, 252, 267, false);
 
     private int x1, x2, y1, y2;
+    private boolean highlightVariant;
 
-    IconContainer(int x1, int x2, int y1, int y2) {
+    IconContainer(int x1, int x2, int y1, int y2, boolean highlightVariant) {
         this.x1 = x1;
         this.x2 = x2;
         this.y1 = y1;
         this.y2 = y2;
+        this.highlightVariant = highlightVariant;
     }
 
     public int getX1() {
@@ -21,7 +24,7 @@ public enum IconContainer {
     }
 
     public int getY1(boolean hovered) {
-        return y1 + (hovered ? getHeight() : 0);
+        return y1 + (hovered && highlightVariant ? getHeight() : 0);
     }
 
     public int getWidth() {
@@ -29,6 +32,6 @@ public enum IconContainer {
     }
 
     public int getHeight() {
-        return (y2 - y1)/2;
+        return highlightVariant ? (y2 - y1)/2 : y2 - y1;
     }
 }
